requirements.txt
crewai>=0.28.0
crewai-tools>=0.1.0
langchain-google-genai>=1.0.0
playwright>=1.40.0
browser-use>=0.1.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
